+++ agent/src/cli/evaluate-autocomplete/AutocompleteMatcher.test.ts
@@ -5,6 +5,7 @@ import dedent from 'dedent'
 import { describe, expect, it } from 'vitest'
 import * as vscode from 'vscode'
 
+import { isNode16 } from '../../index.test'
 import { getLanguageForFileName } from '../../language'
 
 import { AutocompleteMatcher } from './AutocompleteMatcher'
@@ -13,11 +14,6 @@ import { Queries } from './Queries'
 
 function isWindows(): boolean {
     return os.platform().startsWith('win')
-}
-
-function isNode16(): boolean {
-    const [major] = process.versions.node.split('.')
-    return Number.parseInt(major, 10) <= 16
 }
 
 describe.skipIf(isWindows() || isNode16())('AutocompleteMatcher', () => {