+++ lib/ui/src/chat/TranscriptItem.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 
 import classNames from 'classnames'
 
@@ -7,8 +7,6 @@ import type { ChatMessage, Guardrails } from '@sourcegraph/cody-shared'
 import type {
     ApiPostMessage,
     ChatButtonProps,
-    ChatUISubmitButtonProps,
-    ChatUITextAreaProps,
     CodeBlockActionsProps,
     EditButtonProps,
     FeedbackButtonsProps,
@@ -33,7 +31,6 @@ export interface TranscriptItemClassNames {
     codeBlocksCopyButtonClassName?: string
     codeBlocksInsertButtonClassName?: string
     transcriptActionClassName?: string
-    chatInputClassName?: string
 }
 
 /**
@@ -41,22 +38,20 @@ export interface TranscriptItemClassNames {
  */
 export const TranscriptItem: React.FunctionComponent<
     {
+        index: number
         message: ChatMessage
         inProgress: boolean
-        beingEdited: boolean
-        setBeingEdited: (input: boolean) => void
+        beingEdited: number | undefined
+        setBeingEdited: (index?: number) => void
+        EditButtonContainer?: React.FunctionComponent<EditButtonProps>
+        showEditButton: boolean
         fileLinkComponent: React.FunctionComponent<FileLinkProps>
         symbolLinkComponent: React.FunctionComponent<SymbolLinkProps>
-        textAreaComponent?: React.FunctionComponent<ChatUITextAreaProps>
-        EditButtonContainer?: React.FunctionComponent<EditButtonProps>
-        editButtonOnSubmit?: (text: string) => void
-        showEditButton: boolean
         FeedbackButtonsContainer?: React.FunctionComponent<FeedbackButtonsProps>
         feedbackButtonsOnSubmit?: (text: string) => void
         showFeedbackButtons: boolean
         copyButtonOnSubmit?: CodeBlockActionsProps['copyButtonOnSubmit']
         insertButtonOnSubmit?: CodeBlockActionsProps['insertButtonOnSubmit']
-        submitButtonComponent?: React.FunctionComponent<ChatUISubmitButtonProps>
         abortMessageInProgressComponent?: React.FunctionComponent<{
             onAbortMessageInProgress: () => void
         }>
@@ -67,6 +62,7 @@ export const TranscriptItem: React.FunctionComponent<
         guardrails?: Guardrails
     } & TranscriptItemClassNames
 > = React.memo(function TranscriptItemContent({
+    index,
     message,
     inProgress,
     beingEdited,
@@ -79,96 +75,64 @@ export const TranscriptItem: React.FunctionComponent<
     codeBlocksCopyButtonClassName,
     codeBlocksInsertButtonClassName,
     transcriptActionClassName,
-    textAreaComponent: TextArea,
     EditButtonContainer,
-    editButtonOnSubmit,
     showEditButton,
     FeedbackButtonsContainer,
     feedbackButtonsOnSubmit,
     showFeedbackButtons,
     copyButtonOnSubmit,
     insertButtonOnSubmit,
-    submitButtonComponent: SubmitButton,
-    chatInputClassName,
     ChatButtonComponent,
     userInfo,
     postMessage,
     guardrails,
 }) {
-    const [formInput, setFormInput] = useState<string>(message.displayText ?? '')
-    const EditTextArea =
-        TextArea && beingEdited && editButtonOnSubmit && SubmitButton ? (
-            <div className={styles.textAreaContainer}>
-                <TextArea
-                    className={classNames(styles.chatInput, chatInputClassName)}
-                    rows={5}
-                    value={formInput}
-                    autoFocus={true}
-                    required={true}
-                    onInput={event => setFormInput((event.target as HTMLInputElement).value)}
-                    onKeyDown={event => {
-                        if (event.key === 'Escape') {
-                            setBeingEdited(false)
-                        }
-
-                        if (
-                            event.key === 'Enter' &&
-                            !event.shiftKey &&
-                            !event.nativeEvent.isComposing &&
-                            formInput.trim()
-                        ) {
-                            event.preventDefault()
-                            setBeingEdited(false)
-                            editButtonOnSubmit(formInput)
-                        }
-                    }}
-                    chatEnabled={true}
-                />
-                <SubmitButton
-                    className={styles.submitButton}
-                    isFollowUp={false}
-                    onClick={() => {
-                        setBeingEdited(false)
-                        editButtonOnSubmit(formInput)
-                    }}
-                    disabled={formInput.length === 0}
-                />
-            </div>
-        ) : null
+    // A boolean indicating whether the message was sent by a human speaker.
+    const isHumanMessage = message.speaker === 'human'
+    // A boolean that determines if any message is currently being edited.
+    const isInEditingMode = beingEdited !== undefined
+    // A boolean indicating whether the current transcript item is the one being edited.
+    const isItemBeingEdited = beingEdited === index
+    // A boolean indicating whether the current message is a command input.
+    const isCommandInput = message?.displayText?.startsWith('/')
 
     return (
         <div
             className={classNames(
                 styles.row,
                 transcriptItemClassName,
-                message.speaker === 'human' ? humanTranscriptItemClassName : styles.assistantRow
+                isHumanMessage ? humanTranscriptItemClassName : styles.assistantRow,
+                // When editing a message, all other messages (both human and assistant messages) are blurred (unfocused)
+                // except for the current message (transcript item) that is being edited (focused)
+                isInEditingMode && (!isHumanMessage || !isItemBeingEdited) && styles.unfocused,
+                isItemBeingEdited && styles.focused
             )}
         >
-            {showEditButton &&
-                EditButtonContainer &&
-                editButtonOnSubmit &&
-                TextArea &&
-                message.speaker === 'human' && (
-                    <div
-                        className={beingEdited ? styles.editingContainer : styles.editingButtonContainer}
+            {/* Edit button shows up on all human messages, but are hidden during Editing Mode*/}
+            {showEditButton && EditButtonContainer && (
+                <div
+                    className={classNames(
+                        styles.editingButtonContainer,
+                        isInEditingMode && styles.editingButtonContainerIsEditingMode
+                    )}
+                    tabIndex={isInEditingMode ? -1 : undefined}
+                    aria-hidden={isInEditingMode}
+                >
+                    <header
+                        className={classNames(
+                            styles.transcriptItemHeader,
+                            transcriptItemParticipantClassName
+                        )}
                     >
-                        <header
-                            className={classNames(
-                                styles.transcriptItemHeader,
-                                transcriptItemParticipantClassName
-                            )}
-                        >
-                            {beingEdited && (
-                                <p className={classNames(styles.editingLabel)}>Editing...</p>
-                            )}
-                            <EditButtonContainer
-                                className={styles.FeedbackEditButtonsContainer}
-                                messageBeingEdited={beingEdited}
-                                setMessageBeingEdited={setBeingEdited}
-                            />
-                        </header>
-                    </div>
-                )}
+                        <EditButtonContainer
+                            className={styles.feedbackEditButtonsContainer}
+                            messageBeingEdited={index}
+                            setMessageBeingEdited={setBeingEdited}
+                            disabled={isCommandInput || isInEditingMode}
+                        />
+                    </header>
+                </div>
+            )}
             {message.preciseContext && message.preciseContext.length > 0 && (
                 <div className={styles.actions}>
                     <PreciseContexts
@@ -190,23 +154,17 @@ export const TranscriptItem: React.FunctionComponent<
                     />
                 )
             ) : null}
-            <div
-                className={classNames(styles.contentPadding, EditTextArea ? undefined : styles.content)}
-            >
+            <div className={classNames(styles.contentPadding, styles.content)}>
                 {message.displayText ? (
-                    EditTextArea ? (
-                        !inProgress && !message.displayText.startsWith('/') && EditTextArea
-                    ) : (
-                        <CodeBlocks
-                            displayText={message.displayText}
-                            copyButtonClassName={codeBlocksCopyButtonClassName}
-                            copyButtonOnSubmit={copyButtonOnSubmit}
-                            insertButtonClassName={codeBlocksInsertButtonClassName}
-                            insertButtonOnSubmit={insertButtonOnSubmit}
-                            metadata={message.metadata}
-                            guardrails={guardrails}
-                        />
-                    )
+                    <CodeBlocks
+                        displayText={message.displayText}
+                        copyButtonClassName={codeBlocksCopyButtonClassName}
+                        copyButtonOnSubmit={copyButtonOnSubmit}
+                        insertButtonClassName={codeBlocksInsertButtonClassName}
+                        insertButtonOnSubmit={insertButtonOnSubmit}
+                        metadata={message.metadata}
+                        guardrails={guardrails}
+                    />
                 ) : (
                     inProgress && <BlinkingCursor />
                 )}
@@ -214,7 +172,8 @@ export const TranscriptItem: React.FunctionComponent<
             {message.buttons?.length && ChatButtonComponent && (
                 <div className={styles.actions}>{message.buttons.map(ChatButtonComponent)}</div>
             )}
-            {message.speaker === 'human' && (
+            {/* Enhanced Context list shows up on human message only */}
+            {isHumanMessage && (
                 <div className={styles.contextFilesContainer}>
                     {message.contextFiles && message.contextFiles.length > 0 ? (
                         <EnhancedContext
@@ -227,10 +186,11 @@ export const TranscriptItem: React.FunctionComponent<
                     )}
                 </div>
             )}
-            {showFeedbackButtons &&
+            {/* Display feedback buttons on assistant messages only */}
+            {!isHumanMessage &&
+                showFeedbackButtons &&
                 FeedbackButtonsContainer &&
-                feedbackButtonsOnSubmit &&
-                message.speaker === 'assistant' && (
+                feedbackButtonsOnSubmit && (
                     <footer
                         className={classNames(
                             styles.footerContainer,
@@ -238,8 +198,9 @@ export const TranscriptItem: React.FunctionComponent<
                         )}
                     >
                         {/* display edit buttons on last user message, feedback buttons on last assistant message only */}
+                        {/* Hide the feedback buttons during editing mode */}
                         <FeedbackButtonsContainer
-                            className={styles.FeedbackEditButtonsContainer}
+                            className={styles.feedbackEditButtonsContainer}
                             feedbackButtonsOnSubmit={feedbackButtonsOnSubmit}
                         />
                     </footer>